# Conflicts:
#	ProjekatManifestacije/bin/beans/Karta.class
#	ProjekatManifestacije/bin/beans/Kupac.class
#	ProjekatManifestacije/bin/beans/Manifestacija.class
#	ProjekatManifestacije/bin/beans/Prodavac.class
#	ProjekatManifestacije/bin/beans/Tip.class
#	ProjekatManifestacije/bin/dao/AdministratorDAO$1.class
#	ProjekatManifestacije/bin/dao/KartaDAO$1.class
#	ProjekatManifestacije/bin/dao/KartaDAO.class
#	ProjekatManifestacije/bin/dao/KupacDAO.class
#	ProjekatManifestacije/bin/dao/ManifestacijeDAO.class
#	ProjekatManifestacije/bin/dao/ProdavacDAO$1.class
#	ProjekatManifestacije/bin/dao/ProdavacDAO.class
#	ProjekatManifestacije/bin/rest/ProjekatManifestacijeMain.class
#	ProjekatManifestacije/files/manifestacije.json
#	ProjekatManifestacije/files/prodavci.json
#	ProjekatManifestacije/src/beans/Manifestacija.java
#	ProjekatManifestacije/src/beans/Prodavac.java
#	ProjekatManifestacije/src/dao/KartaDAO.java
#	ProjekatManifestacije/src/dao/ManifestacijeDAO.java
#	ProjekatManifestacije/src/rest/ProjekatManifestacijeMain.java
#	ProjekatManifestacije/static/index.html

+++ ProjekatManifestacije/src/beans/Lokacija.java
@@ -2,8 +2,8 @@ package beans;
 
 public class Lokacija {
 	private int id;
-	private int geografskaDuzina;
-	private int geografskaSirina;
+	private double geografskaDuzina;
+	private double geografskaSirina;
 	private Adresa adresa;
 	private boolean obrisana;
 	
@@ -17,19 +17,19 @@ public class Lokacija {
 		this.id = id;
 	}
 
-	public int getGeografskaDuzina() {
+	public double getGeografskaDuzina() {
 		return geografskaDuzina;
 	}
 
-	public void setGeografskaDuzina(int geografskaDuzina) {
+	public void setGeografskaDuzina(double geografskaDuzina) {
 		this.geografskaDuzina = geografskaDuzina;
 	}
 
-	public int getGeografskaSirina() {
+	public double getGeografskaSirina() {
 		return geografskaSirina;
 	}
 
-	public void setGeografskaSirina(int geografskaSirina) {
+	public void setGeografskaSirina(double geografskaSirina) {
 		this.geografskaSirina = geografskaSirina;
 	}
